# storage/api-docs/api-docs-admin-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Management System API",
        "description": "API documentation for the School Management System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        }
    ],
    "paths": {
...
        "/admin/class": {
            "post": {
                "tags": [
                    "Class Management"
                ],
                "summary": "Create a new Class",
                "description": "Add a new class to the system.",
                "operationId": "f4c9c31d35733dd02547d1c0c5080fb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "class_grade",
                                    "section_number",
                                    "capacity",
                                    "location"
                                ],
                                "properties": {
                                    "class_grade": {
                                        "description": "Grade level of the class",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Grade 10"
                                    },
                                    "section_number": {
                                        "description": "Section number of the class",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "A"
                                    },
                                    "capacity": {
                                        "description": "Maximum capacity of the class",
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "location": {
                                        "description": "Location of the class",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Room 101"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Class created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Class created successfully"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/class-course": {
            "post": {
                "tags": [
                    "Class Management"
                ],
                "summary": "Add a course to a Class",
                "description": "Add a course to a class.",
                "operationId": "81a02eab6a707a5e7f39845718cffa49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "class_id",
                                    "course_id"
                                ],
                                "properties": {
                                    "class_id": {
                                        "description": "Class id that you want to add the course to.",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "course_id": {
                                        "description": "The added course.",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Class created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Class created successfully"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/class/{id}": {
            "put": {
                "tags": [
                    "Class Management"
                ],
                "summary": "Update an Existing Class",
                "description": "Update the details of an existing class.",
                "operationId": "76fefd6d4fdfa2ed15677ce8c665fad1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the class to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "class_grade",
                                    "section_number",
                                    "capacity",
                                    "location"
                                ],
                                "properties": {
                                    "class_grade": {
                                        "description": "Grade level of the class",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Grade 11"
                                    },
                                    "section_number": {
                                        "description": "Section number of the class",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "B"
                                    },
                                    "capacity": {
                                        "description": "Maximum capacity of the class",
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "location": {
                                        "description": "Location of the class",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Room 202"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Class updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Class updated successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Class Management"
                ],
                "summary": "Delete a Class",
                "description": "Delete an existing class.",
                "operationId": "abac60653d876e118c36fca7f1b6a523",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the class to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Class deleted successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}